+++ src/pages/inventory/modals/Additem/components/Additem.tsx
@@ -9,7 +9,7 @@ import {
   Button,
   useToast,
 } from "@chakra-ui/react";
-import { useAddInventoryForm } from "@/pages/inventory/hooks/useAddInventoryForm";
+import { useAddInventoryForm } from "@/pages/inventory/modals/Additem/hooks/useAddInventoryForm";
 import { AddInventoryForm } from "./AddInventoryForm";
 import { ExistingProductAlert } from "./ExistingProductAlert";
