+++ src/pages/inventory/modals/BarcodeScanner/components/ManualPane.tsx
