+++ src/pages/inventory/modals/BarcodeScanner/components/ScannerPane.tsx
