+++ src/pages/inventory/modals/BarcodeScanner/hooks/useBarcodeScanner.tsx
