+++ src/pages/inventory/modals/BarcodeScanner/index.ts
@@ -1,6 +1,6 @@
-export { BarcodeScanner } from "./BarcodeScanner";
-export type { BarcodeScannerProps } from "./BarcodeScanner";
-export { ScannerPane } from "./ScannerPane";
-export type { ScannerPaneProps } from "./ScannerPane";
-export { ManualPane } from "./ManualPane";
-export type { ManualPaneProps } from "./ManualPane";
+export { BarcodeScanner } from "./components/BarcodeScanner";
+export type { BarcodeScannerProps } from "./components/BarcodeScanner";
+export { ScannerPane } from "./components/ScannerPane";
+export type { ScannerPaneProps } from "./components/ScannerPane";
+export { ManualPane } from "./components/ManualPane";
+export type { ManualPaneProps } from "./components/ManualPane";
