+++ src/pages/inventory/modals/ItemDetails/components/DeleteConfirmationDialog.tsx
