+++ src/pages/inventory/modals/ItemDetails/components/InventoryDisplay.tsx
