+++ src/pages/inventory/modals/ItemDetails/components/InventoryHeader.tsx
